# big2/game_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Big 2 Game Server",
    "description": "HTTP API for Big 2 card game with JSON request/response",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/game/register/{game_id}": {
      "post": {
        "tags": [
          "Game Management"
        ],
        "summary": "Register Game",
        "description": "Register a new game.\n\nCreates a new game with the specified ID. The game will be in 'waiting' state\nuntil 4 players join.",
        "operationId": "register_game_api_game_register__game_id__post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique game identifier",
              "title": "Game Id"
            },
            "description": "Unique game identifier"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterGameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Game already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/join/{game_id}": {
      "post": {
        "tags": [
          "Game Management"
        ],
        "summary": "Join Game",
        "description": "Join an existing game.\n\nAdds a player to the game. When the 4th player joins, the game automatically\nstarts and cards are dealt.",
        "operationId": "join_game_api_game_join__game_id__post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Game ID to join",
              "title": "Game Id"
            },
            "description": "Game ID to join"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully joined game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinGameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot join game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/state/{game_id}": {
      "get": {
        "tags": [
          "Game State"
        ],
        "summary": "Get Game State",
        "description": "Get current game state.\n\nReturns complete game state including turn information, move history,\ncards remaining per player, and winner (if game finished).",
        "operationId": "get_game_state_api_game_state__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Game ID",
              "title": "Game Id"
            },
            "description": "Game ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game state retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGameStateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/action/{game_id}": {
      "post": {
        "tags": [
          "Game Actions"
        ],
        "summary": "Submit Action",
        "description": "Submit a game action (play cards or pass).\n\nValidates the move according to Big 2 rules and applies it if legal.\nReturns detailed status including success/failure reason.\n\nThe action is validated for:\n- Correct player turn\n- Valid card combination\n- Legal move according to Big 2 rules\n- Player owns the cards being played",
        "operationId": "submit_action_api_game_action__game_id__post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Game ID",
              "title": "Game Id"
            },
            "description": "Game ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action processed (check status field for result)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/games": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Games",
        "description": "List all game IDs (admin/debug endpoint).",
        "operationId": "list_games_api_admin_games_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Games Api Admin Games Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/game/{game_id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete Game",
        "description": "Delete a game (admin/debug endpoint).",
        "operationId": "delete_game_api_admin_game__game_id__delete",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Game ID to delete",
              "title": "Game Id"
            },
            "description": "Game ID to delete"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionStatus": {
        "type": "string",
        "enum": [
          "success",
          "error",
          "illegal_move",
          "out_of_turn",
          "stale_turn",
          "game_over"
        ],
        "title": "ActionStatus",
        "description": "Action response status."
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "play",
          "pass"
        ],
        "title": "ActionType",
        "description": "Action types."
      },
      "Card": {
        "properties": {
          "suit": {
            "$ref": "#/components/schemas/Suit",
            "description": "The suit of the card"
          },
          "rank": {
            "$ref": "#/components/schemas/Rank",
            "description": "The rank of the card"
          }
        },
        "type": "object",
        "required": [
          "suit",
          "rank"
        ],
        "title": "Card",
        "description": "A playing card.",
        "examples": [
          {
            "rank": "10",
            "suit": "hearts"
          },
          {
            "rank": "A",
            "suit": "spades"
          }
        ]
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Generic error response.",
        "examples": [
          {
            "detail": "Game not found"
          }
        ]
      },
      "GameState": {
        "type": "string",
        "enum": [
          "waiting",
          "started",
          "finished",
          "error"
        ],
        "title": "GameState",
        "description": "Game state enum."
      },
      "GetGameStateResponse": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/GameState",
            "description": "Current game state"
          },
          "turn_number": {
            "type": "integer",
            "title": "Turn Number",
            "description": "Monotonically increasing turn sequence"
          },
          "turn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn",
            "description": "Current player ID (null if game not started)"
          },
          "turn_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Turn Order",
            "description": "Fixed turn order (set when game starts)"
          },
          "last_move_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Move Id",
            "description": "Server-generated ID of last move"
          },
          "last_move_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Move At",
            "description": "Timestamp of last move"
          },
          "move_count": {
            "type": "integer",
            "title": "Move Count",
            "description": "Total applied moves for this game"
          },
          "cards_left": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cards Left",
            "description": "Number of cards remaining per player"
          },
          "game_history": {
            "items": {
              "$ref": "#/components/schemas/MoveHistoryEntry"
            },
            "type": "array",
            "title": "Game History",
            "description": "Complete game history"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner",
            "description": "Winner player ID (if game finished)"
          }
        },
        "type": "object",
        "required": [
          "state",
          "turn_number",
          "move_count"
        ],
        "title": "GetGameStateResponse",
        "description": "Response for getting game state.",
        "examples": [
          {
            "cards_left": {
              "player_1": 10,
              "player_2": 8,
              "player_3": 5,
              "player_4": 12
            },
            "game_history": [
              {
                "cards": [
                  {
                    "rank": "10",
                    "suit": "hearts"
                  },
                  {
                    "rank": "10",
                    "suit": "diamonds"
                  }
                ],
                "player": "player_1"
              }
            ],
            "last_move_at": "2021-01-01T00:05:02Z",
            "last_move_id": "bcaa9f25-a8ef-4db1-a833-7f8dba0c9f3d",
            "move_count": 18,
            "state": "started",
            "turn": "player_1",
            "turn_number": 12,
            "turn_order": [
              "player_1",
              "player_2",
              "player_3",
              "player_4"
            ]
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JoinGameResponse": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id",
            "description": "Unique game identifier"
          },
          "game_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Game Created At",
            "description": "ISO 8601 timestamp of game creation"
          },
          "player_id": {
            "type": "string",
            "title": "Player Id",
            "description": "Unique player identifier"
          },
          "num_players": {
            "type": "integer",
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Num Players",
            "description": "Number of players currently in the game"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Secret token for authentication"
          },
          "state": {
            "$ref": "#/components/schemas/GameState",
            "description": "Current game state"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "game_created_at",
          "player_id",
          "num_players",
          "token",
          "state"
        ],
        "title": "JoinGameResponse",
        "description": "Response for joining a game.",
        "examples": [
          {
            "game_created_at": "2021-01-01T00:00:00Z",
            "game_id": "1234567890",
            "num_players": 2,
            "player_id": "player_abc",
            "state": "waiting",
            "token": "secret_token_xyz"
          }
        ]
      },
      "MoveHistoryEntry": {
        "properties": {
          "player": {
            "type": "string",
            "title": "Player",
            "description": "Player ID who made the move"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "type": "array",
            "title": "Cards",
            "description": "Cards played (empty array means pass)"
          }
        },
        "type": "object",
        "required": [
          "player",
          "cards"
        ],
        "title": "MoveHistoryEntry",
        "description": "A single move in game history.",
        "examples": [
          {
            "cards": [
              {
                "rank": "10",
                "suit": "hearts"
              },
              {
                "rank": "10",
                "suit": "diamonds"
              }
            ],
            "player": "player_1"
          },
          {
            "cards": [],
            "player": "player_2"
          }
        ]
      },
      "Rank": {
        "type": "string",
        "enum": [
          "3",
          "4",
          "5",
          "6",
          "7",
          "8",
          "9",
          "10",
          "J",
          "Q",
          "K",
          "A",
          "2"
        ],
        "title": "Rank",
        "description": "Card ranks."
      },
      "RegisterGameResponse": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id",
            "description": "Unique game identifier"
          },
          "game_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Game Created At",
            "description": "ISO 8601 timestamp of game creation"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "game_created_at"
        ],
        "title": "RegisterGameResponse",
        "description": "Response for game registration.",
        "examples": [
          {
            "game_created_at": "2021-01-01T00:00:00Z",
            "game_id": "1234567890"
          }
        ]
      },
      "SubmitActionRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionType",
            "description": "Action type"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "type": "array",
            "title": "Cards",
            "description": "Cards to play (empty for pass)"
          },
          "expected_turn_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Turn Number",
            "description": "Optional guard against stale/out-of-order moves"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Secret authentication token"
          }
        },
        "type": "object",
        "required": [
          "action",
          "token"
        ],
        "title": "SubmitActionRequest",
        "description": "Request to submit an action (play or pass).",
        "examples": [
          {
            "action": "play",
            "cards": [
              {
                "rank": "10",
                "suit": "hearts"
              },
              {
                "rank": "10",
                "suit": "diamonds"
              }
            ],
            "expected_turn_number": 12,
            "token": "secret_token_xyz"
          },
          {
            "action": "pass",
            "cards": [],
            "token": "secret_token_xyz"
          }
        ]
      },
      "SubmitActionResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ActionStatus",
            "description": "Result status"
          },
          "applied_move_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Move Id",
            "description": "Server-generated move ID (if applied)"
          },
          "turn_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Number",
            "description": "Next expected turn number (if applied)"
          },
          "move_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Count",
            "description": "Total move count after this action"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Error/rejection reason (present on non-success)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SubmitActionResponse",
        "description": "Response to action submission.",
        "examples": [
          {
            "applied_move_id": "bcaa9f25-a8ef-4db1-a833-7f8dba0c9f3d",
            "move_count": 19,
            "status": "success",
            "turn_number": 13
          },
          {
            "reason": "expected turn_number 12 but got 11",
            "status": "stale_turn"
          }
        ]
      },
      "Suit": {
        "type": "string",
        "enum": [
          "diamonds",
          "clubs",
          "hearts",
          "spades"
        ],
        "title": "Suit",
        "description": "Card suits."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}